refactor(WaitingController): 예약대기 컨트롤러 분리
